hello: Bring back the count size to u8
This shows the behavior of the wrap around,
which just wraps around and the data account
doesn't care about it.  It's all up to how
the program behaves.

+++ hello/src/lib.rs
@@ -19,7 +19,7 @@ use solana_program::{
 #[derive(Debug, BorshSerialize, BorshDeserialize)]
 pub struct Counter {
     /// Number of greetings it received.
-    pub count: u16,
+    pub count: u8,
 }
 
 // Declares and export the program's entrypoint.
